requirements.txt
djangorestframework
langchain
langchain-google-genai
google-generativeai
python-dotenv
